+++ database_SQL/blog_DB_data.sql
@@ -1,22 +1,13 @@
-SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
-START TRANSACTION;
-SET time_zone = "+00:00";
-
-USE `blog`;
-
---
--- Inserting data into table `admin`
---
-
+-- `admin`
 INSERT INTO `admin` (`id`, `login`, `password`) VALUES
-(1, 'admin', '$2y$10$L9fQlnPTTuYkNhLnXh68..F8R.bJdLaAJBJjXU8RjhiHUHFVVJyCe');
+(1, 'admin', '$2y$10$w9GkWVGXxSTjw4A9QyjasuqyeJyUPp2JlWYnFGFEen7e2..YUXxNC');
 
---
--- Inserting data into table `service`
---
-
+-- `service`
 INSERT INTO `service` (`id`, `service_status`) VALUES
-(1, 1);
+(1, 0);
 
+-- `settings`
+INSERT INTO `settings` (`id`, `name`, `value`) VALUES
+(1, 'active_theme', 'purple-show');
 
 COMMIT;