+++ public/handlers/connect.php
@@ -3,7 +3,7 @@
 
     $host = "localhost";
     $db_user = "root";
-    $db_password ="";
+    $db_password ="root";
     $db_name = "blog";