+++ public/handlers/index.php
@@ -1,0 +1,4 @@
+<?php // Secruity 
+
+header('Location: ../index.php');
+exit();