+++ public/single.php
@@ -3,10 +3,10 @@
 
 session_start();
 
-include_once 'functions.php';
-require_once 'service.alg.php';
-include_once 'single.alg.php';
-include 'connect.php';
+include './handlers/functions.php';
+require_once './handlers/service.alg.php';
+require_once './handlers/cookies.php';
+include_once './handlers/single.alg.php';
 
 service();
 
@@ -15,6 +15,9 @@ if (!isset($_GET['title']) || empty(trim($_GET['title'])) || !preg_match('/^[a-z
     header("Location: error_404.php");
     exit();
 }
+
+$page = 'single.php?title='.$_GET['title'];
+cookie($page);
 
 $restored_title = str_replace('_', ' ', filter_var($_GET['title'], FILTER_SANITIZE_SPECIAL_CHARS)); //optionally use - insted of _
 
@@ -25,9 +28,13 @@ $restored_title = str_replace('_', ' ', filter_var($_GET['title'], FILTER_SANITI
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title><?php echo htmlspecialchars($restored_title); ?></title> <!-- XSS prevention -->
 
-    <link rel="stylesheet" href="main.css" />
-    <link rel="stylesheet" href="single.css" />
+    <?php
+        include "../themes/index.php";
+    ?>
 
+    <!-- == SYSTEM == -->
+    <link rel="stylesheet" href="./CSS/popout.css" />
+    <script src="./JS/popout.js"></script>
 </head>
 <body>
     <header>
@@ -57,6 +64,10 @@ $restored_title = str_replace('_', ' ', filter_var($_GET['title'], FILTER_SANITI
             foot();
         ?>
     </footer>
+
+    <?php
+        cookie_popout()
+    ?>
 
 </body>
 </html>