+++ themes/handlers/connect.php
@@ -1,3 +1,4 @@
+
 <?php
 
     $host = "localhost";
@@ -6,6 +7,3 @@
     $db_name = "blog";
 
     
-
-
-