+++ CHANGELOG.md
@@ -2,18 +2,37 @@
 
 All notable changes to this project will be documented in this file.
 
-## v1.0.0 – First Release
+## New in v1.2.0
 
-### New in v1.0.0
-
-The project is ready for use but is still evolving. Work is ongoing to fix bugs and address security issues.
+Changes since 1.1.0
 
 ### Added
 <!-- - Initial examples of added changelog entries. -->
+
+`/root`
+1 - themes folder
+
+`/public`
+1 - cookies for users,
+2 - tables for cookies added
+3 - JS folder
+4 - popout for cookies
+5 - collecting data for analytics
+
+`/admin`
+1 - added 2 new windows: theme and analytics
+2 - build algo to change theme in db
 .....
 
 ### Changed
 <!-- - Improved existing changelog entries for consistency. -->
+
+`/public`
+1 - structure of /public fixes
+
+`/admin`
+1 - chagne name of setting window,function to password
+2 - error function
 .....
 
 ### Fixed